:recycle: api: move get-next-challenge-for-certification dependencies to src
