+++ api/tests/certification/flash-certification/integration/application/scenario-simulator-controller_test.js
@@ -4,7 +4,7 @@ import * as moduleUnderTest from '../../../../../src/certification/flash-certifi
 import { random } from '../../../../../lib/infrastructure/utils/random.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { pickAnswerStatusService } from '../../../../../lib/domain/services/pick-answer-status-service.js';
-import { pickChallengeService } from '../../../../../lib/domain/services/pick-challenge-service.js';
+import { pickChallengeService } from '../../../../../src/certification/flash-certification/domain/services/pick-challenge-service.js';
 
 describe('Integration | Application | scenario-simulator-controller', function () {
   let httpTestServer;
