+++ api/tests/certification/flash-certification/unit/domain/services/pick-challenge-service_test.js
@@ -1,6 +1,6 @@
-import { expect, domainBuilder } from '../../../test-helper.js';
-import { pickChallengeService } from '../../../../lib/domain/services/pick-challenge-service.js';
-import { LOCALE } from '../../../../src/shared/domain/constants.js';
+import { expect, domainBuilder } from '../../../../../test-helper.js';
+import { pickChallengeService } from '../../../../../../src/certification/flash-certification/domain/services/pick-challenge-service.js';
+import { LOCALE } from '../../../../../../src/shared/domain/constants.js';
 
 const { ENGLISH_SPOKEN, FRENCH_FRANCE, FRENCH_SPOKEN } = LOCALE;
