+++ api/tests/certification/shared/integration/infrastructure/repositories/certification-challenge-repository_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
-import { expect, domainBuilder, databaseBuilder } from '../../../test-helper.js';
-import { CertificationChallenge } from '../../../../lib/domain/models/CertificationChallenge.js';
-import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
-import * as certificationChallengeRepository from '../../../../lib/infrastructure/repositories/certification-challenge-repository.js';
-import { CertificationVersion } from '../../../../src/shared/domain/models/CertificationVersion.js';
+import { expect, domainBuilder, databaseBuilder } from '../../../../../test-helper.js';
+import { CertificationChallenge } from '../../../../../../lib/domain/models/CertificationChallenge.js';
+import { AssessmentEndedError } from '../../../../../../lib/domain/errors.js';
+import * as certificationChallengeRepository from '../../../../../../src/certification/shared/infrastructure/repositories/certification-challenge-repository.js';
+import { CertificationVersion } from '../../../../../../src/shared/domain/models/CertificationVersion.js';
 
 describe('Integration | Repository | Certification Challenge', function () {
   describe('#save', function () {
