+++ api/tests/shared/integration/infrastructure/repositories/challenge-repository_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
-import { catchErr, domainBuilder, expect, mockLearningContent } from '../../../test-helper.js';
-import { Challenge } from '../../../../src/shared/domain/models/Challenge.js';
-import { Validator } from '../../../../lib/domain/models/Validator.js';
-import * as challengeRepository from '../../../../src/shared/infrastructure/repositories/challenge-repository.js';
-import { NotFoundError } from '../../../../lib/domain/errors.js';
+import { catchErr, domainBuilder, expect, mockLearningContent } from '../../../../test-helper.js';
+import { Challenge } from '../../../../../src/shared/domain/models/Challenge.js';
+import { Validator } from '../../../../../lib/domain/models/Validator.js';
+import * as challengeRepository from '../../../../../src/shared/infrastructure/repositories/challenge-repository.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
 
 describe('Integration | Repository | challenge-repository', function () {
   describe('#get', function () {
